i have a problem

+++ web-library/src/main/java/ru/skypro/lessons/springboot/weblibrary/pojo/ReportRepository.java
@@ -2,6 +2,6 @@ package ru.skypro.lessons.springboot.weblibrary.pojo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
     Report findByEmployee(Employee employee);
 }
